TAO-3697 Removed usages of old filesystem and deleted unused classes
