+++ models/classes/websource/BaseWebsource.php
@@ -22,7 +22,6 @@
 namespace oat\tao\model\websource;
 
 use oat\oatbox\Configurable;
-use core_kernel_fileSystem_FileSystem;
 use oat\oatbox\filesystem\FileSystemService;
 use oat\oatbox\service\ServiceManager;
 use League\Flysystem\FileNotFoundException;
@@ -46,8 +45,6 @@ implements Websource
     
 	/**
 	 * Filesystem that is being made available
-	 * 
-	 * @var core_kernel_fileSystem_FileSystem
 	 */
 	protected $fileSystem = null;
