+++ models/classes/websource/DirectWebSource.php
@@ -20,9 +20,6 @@
  */
 namespace oat\tao\model\websource;
 
-use core_kernel_fileSystem_FileSystem;
-use common_ext_ExtensionsManager;
-
 /**
  * Grants Access to compiled data via a direct URL link
  *
